refactor(ui/QInput): move mask api tests to own suite
